--- a/turbopack/crates/turbo-tasks-memory/tests/detached.rs
+++ b/turbopack/crates/turbo-tasks-memory/tests/detached.rs
@@ -1,0 +1,1 @@
+../../turbo-tasks-testing/tests/detached.rs